requirements.txt
tqdm
pandas
langchain
streamlit
tiktoken
python-dotenv
ffmpeg-python
openai-whisper
openai==0.27.8
pinecone-client
langchain_community
youtube-transcript-api
git+https://github.com/openai/whisper.git 


  






